+++ src/data/articleInfo.json
@@ -9,8 +9,8 @@
     "author" : {
         "name" : "hyunJu",
         "image" : "../logo.svg",
-        "github" : "lxxjn0",
-        "blog" : "https://ell.com",
+        "github" : "https://naver.com",
+        "blog" : "https://naver.com",
         "description" : "London 안녕하세요 is the capital city of England. It is the most populous city in the United Kingdom, with a metropolitan area of over 13 million inhabitants.Standing on the River Thames, London has been a major settlement for two millennia, its history going back to its founding by the Romans, who named it Londinium."
     },
     "createdDate" : "20200411",
